outlines sharding optimizations
